move utils to folder and extract labels function to dedicated file
