+++ src/acl/access-controller.ts
@@ -9,7 +9,7 @@ import { AccessGroups } from './access-groups'
 import { DataBrowserContext } from 'pane-registry'
 import { shortNameForFolder } from './acl-control'
 import { currentUser } from '../authn/authn'
-import * as utils from '../utils.js'
+import * as utils from '../utils'
 import * as debug from '../debug'
 import ns from '../ns'
