+++ src/acl/acl-control.ts
@@ -6,7 +6,7 @@
  */
 
 import * as ns from '../ns'
-import * as utils from '../utils.js'
+import * as utils from '../utils'
 import { getACLorDefault, getProspectiveHolder } from './acl'
 import { IndexedFormula, NamedNode } from 'rdflib'
 import { DataBrowserContext } from 'pane-registry'
