+++ src/authn/authn.ts
@@ -25,7 +25,7 @@ import { PaneDefinition } from 'pane-registry'
 import { Signup } from './signup'
 import * as widgets from '../widgets'
 import * as ns from '../ns.js'
-import * as utils from '../utils.js'
+import * as utils from '../utils'
 import { alert } from '../log'
 import { AppDetails, AuthenticationContext } from './types'
 import * as debug from '../debug'
